requirements.txt
pandas
pandas_ta
pybit
python-dotenv